refactor: group into PaymentProvidersAccordion
